Sync interfaces and move Player definition
