+++ src/interfaces/game.go
@@ -2,8 +2,9 @@ package interfaces
 
 //go:generate mockgen --destination=./mocks/table.go interfaces Game
 type Game interface {
+	StartGame()
 	TakeTurn()
-	Accepted() bool
+	IsAccepted() bool
 	GameFinished() bool
-	Channel() string
+	GetChannel() string
 }
